test(resolve): move extension-resolution tests to a sibling file (2000-line gate)
The new tests pushed `resolve/tests.rs` to 2028 lines, tripping the
`check_file_size.sh` 2000-line lint gate. Relocate them to a dedicated
`resolve/extension_resolution_tests.rs` module (mirroring `bun_store_tests`),
bringing `tests.rs` back to 1939 lines. No test changes.
